Fix Play Console 'Unsafe TrustManager / HostnameVerifier' warnings
Remove the unused MyApplication.java (a global trust-all X509TrustManager plus allow-all HostnameVerifier). It was dead code (the manifest registers the default application class) but shipped in the DEX and tripped the Play Console scan. Replace flutter_downloader (whose DownloadWorker bundles the same trust-all) with background_downloader, rewriting DownloadManager to use FileDownloader().updates and dropping the isolate/SendPort plumbing. The app never passed flutter_downloader ignoreSsl, so there is no functional loss. The Dart HttpOverrides self-signed bypass is intentionally left as-is (separate follow-up; not flagged by Play).

+++ lib/objects/download_tracker.dart
@@ -4,7 +4,8 @@ class DownloadTracker {
   String serverUrl;
   String filePath;
 
-  int progress = 0;
+  // 0.0–1.0 (background_downloader progress scale).
+  double progress = 0.0;
 
   // These can be set to update downlaod progress for a particular item
   // you should always check if these exist before using them
